refactor OpticalGater
emulate.FileOpticalGater works
cli.PiOpticalGater not tested
