+++ open_optical_gating/cli/emulate.py
@@ -9,24 +9,24 @@ from loguru import logger
 from skimage import io
 
 # Local imports
-import open_optical_gating.cli.cli as cli
+import open_optical_gating.cli.optical_gater_server as server
 
 
-class OpticalGater(cli.OpticalGater):
-    """Custom class to extend optical gater for emulating from a data file.
+class FileOpticalGater(server.OpticalGater):
+    """Extends the optical gater server for a data file.
     """
 
     def __init__(
-        self, file=None, settings=None, ref_frames=None, ref_frame_period=None
+        self, source=None, settings=None, ref_frames=None, ref_frame_period=None
     ):
         """Function inputs:
-            camera - the raspberry picam PiCamera object
+            source - the raspberry picam PiCamera object
             settings - a dictionary of settings (see default_settings.json)
         """
 
         # initialise parent
-        super(OpticalGater, self).__init__(
-            camera=file,
+        super(FileOpticalGater, self).__init__(
+            source=source,
             settings=settings,
             ref_frames=ref_frames,
             ref_frame_period=ref_frame_period,
@@ -44,11 +44,6 @@ class OpticalGater(cli.OpticalGater):
         self.width, self.height = self.data[0].shape
         self.framerate = self.settings["brightfield_framerate"]
 
-    def init_hardware(self):
-        """As this is an emulator, this function just outputs a log and returns a success."""
-        logger.success("Hardware would be initialised now.")
-        return 0  # default return
-
     def next_frame(self):
         """This function gets the next frame from the data source, which can be passed to analyze()"""
         self.emulate_frame = self.emulate_frame + 1
@@ -57,15 +52,11 @@ class OpticalGater(cli.OpticalGater):
             self.stop = True
         return self.data[self.emulate_frame, :, :]
 
-    def trigger_fluorescence_image_capture(self, delay):
-        """As this is an emulator, this function just outputs a log that a trigger would have been sent."""
-        logger.success("A fluorescence image would be triggered now.")
-
 
 def run(settings):
     """Emulated data capture for a set of sample brightfield frames."""
     logger.success("Initialising gater...")
-    analyser = OpticalGater(file=settings["path"], settings=settings,)
+    analyser = FileOpticalGater(source=settings["path"], settings=settings,)
 
     logger.success("Determining reference period...")
     while analyser.state > 0:
